+++ changelog.md
@@ -4,10 +4,12 @@
 
   * Add instances for `Solo` (GHC-9)
     ([#69](https://github.com/haskell/deepseq/pull/69))
+  * Add once again `infixr 0 deepseq`
+    ([#56](https://github.com/haskell/deepseq/pull/56), [#74](https://github.com/haskell/deepseq/issues/74))
 
 ## 1.4.6.1
 
-  * Revert `infixr 0 deepseq`; this is a breaking change and requires a major version bump
+  * Revert `infixr 0 deepseq`; this does not appear in the version of `deepseq` pinned to GHC 9.2.1
 
 ## 1.4.6.0